refactor: moves coin list into features folder
