+++ src/components/Header.test.tsx
@@ -2,9 +2,9 @@ import { vi, describe, it } from 'vitest';
 import { fireEvent, screen, waitFor } from '@testing-library/react';
 import Header from './Header';
 import { render } from '../test/testUtils';
-
+import { CoinPrice } from '../features/coin-list/types/CoinPrice';
 import * as newsQueryClient from '../features/news/api/getNews';
-import * as coinQueryClient from './api/coins';
+import * as coinQueryClient from '../features/coin-list/api/coins';
 
 describe('Header component', () => {
     it('toggling between favourites change full list to active', () => {
@@ -23,7 +23,7 @@ describe('Header component', () => {
     });
 
     it('news and coin price is refetched when refresh control is clicked', () => {
-        const fetchCoinPriceSpy = vi.spyOn(coinQueryClient, 'fetchCoinPrice').mockResolvedValueOnce({} as coinQueryClient.CoinPrice);
+        const fetchCoinPriceSpy = vi.spyOn(coinQueryClient, 'fetchCoinPrice').mockResolvedValueOnce({} as CoinPrice);
         const fetchNewsSpy = vi.spyOn(newsQueryClient, 'getNews').mockResolvedValueOnce(null);
         render(<Header />);
         fireEvent.click(screen.getByTestId('sync'));
