+++ src/features/coin-list/types/Coin.ts
@@ -1,0 +1,6 @@
+export interface Coin {
+    id: string;
+    code: string;
+    name: string;
+    imageURL: string;
+}
