+++ src/features/coin-list/types/CoinPrice.ts
@@ -1,0 +1,4 @@
+export interface CoinPrice {
+    GBP: string;
+    USD: string;
+}
