+++ src/features/coin-list/types/CoinPriceResponse.ts
@@ -1,0 +1,3 @@
+import { CoinPrice } from './CoinPrice';
+
+export type CoinPriceResponse = { [key: string]: CoinPrice };
